Extract logic, break up functions into separate domain directories and files
