+++ tests/automated/integration/api/review/test_next_source.py
@@ -1,7 +1,7 @@
 import pytest
 
 from src.core.enums import SuggestedStatus, RecordType
-from tests.helpers.complex_test_data_functions import setup_for_get_next_url_for_final_review
+from tests.helpers.setup.final_review.core import setup_for_get_next_url_for_final_review
 
 
 @pytest.mark.asyncio
