+++ tests/automated/integration/db/client/annotate_url/test_agency_not_in_db.py
@@ -2,7 +2,7 @@ import pytest
 
 from src.db.constants import PLACEHOLDER_AGENCY_NAME
 from src.db.models.instantiations.agency import Agency
-from tests.helpers.complex_test_data_functions import setup_for_annotate_agency
+from tests.helpers.setup.annotate_agency.core import setup_for_annotate_agency
 from tests.helpers.db_data_creator import DBDataCreator
