+++ tests/automated/integration/db/client/annotate_url/test_marked_not_relevant.py
@@ -2,7 +2,7 @@ import pytest
 
 from src.core.enums import SuggestedStatus
 from src.db.dtos.url.mapping import URLMapping
-from tests.helpers.complex_test_data_functions import setup_for_get_next_url_for_annotation
+from tests.helpers.setup.annotation.core import setup_for_get_next_url_for_annotation
 from tests.helpers.db_data_creator import DBDataCreator
