+++ tests/automated/integration/db/client/get_next_url_for_final_review/test_batch_id_filtering.py
@@ -1,6 +1,6 @@
 import pytest
 
-from tests.helpers.complex_test_data_functions import setup_for_get_next_url_for_final_review
+from tests.helpers.setup.final_review.core import setup_for_get_next_url_for_final_review
 from tests.helpers.db_data_creator import DBDataCreator
