+++ tests/automated/integration/db/client/get_next_url_for_final_review/test_favor_more_components.py
@@ -1,7 +1,7 @@
 import pytest
 
 from src.core.enums import SuggestionType
-from tests.helpers.complex_test_data_functions import setup_for_get_next_url_for_final_review
+from tests.helpers.setup.final_review.core import setup_for_get_next_url_for_final_review
 from tests.helpers.db_data_creator import DBDataCreator
