+++ tests/automated/integration/db/client/get_next_url_for_user_relevance_annotation/test_pending.py
@@ -1,7 +1,7 @@
 import pytest
 
 from src.core.enums import SuggestedStatus
-from tests.helpers.complex_test_data_functions import setup_for_get_next_url_for_annotation
+from tests.helpers.setup.annotation.core import setup_for_get_next_url_for_annotation
 from tests.helpers.db_data_creator import DBDataCreator
