+++ tests/helpers/setup/annotate_agency/model.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class AnnotateAgencySetupInfo(BaseModel):
+    batch_id: int
+    url_ids: list[int]
